refactor (frontend): remove unused components, move PageWrapper to components/common/
