+++ frontend/src/extra/AnimatedSheet.tsx
