+++ frontend/src/pages/ApiKeys.tsx
@@ -6,7 +6,7 @@ import { useAuth } from "@clerk/clerk-react"
 import { deleteApiKey, getApiKeys } from "@/utils/apiKey.utils"
 import toast from "react-hot-toast"
 import { convertMongoTimestampToLocal } from "@/utils/convertMongoTimestampToLocal"
-import PageWrapper from "@/components/PageWrapper"
+import PageWrapper from "@/components/common/PageWrapper"
 import ApiCreateDialog from "@/components/api-keys/ApiCreateDialog"
 import ApiKeySkeleton from "@/components/api-keys/ApiKeySkeleton"
 import { ReactNode } from "react"
